--- a/lib/apply.js
+++ b/lib/apply.js
@@ -45,7 +45,7 @@ import slice from './internal/slice';
  * two
  * three
  */
-export default function(fn/*, args*/) {
+export default function(fn/*, ...args*/) {
     var args = slice(arguments, 1);
     return function(/*callArgs*/) {
         var callArgs = slice(arguments);