--- a/lib/auto.js
+++ b/lib/auto.js
@@ -4,8 +4,8 @@ import indexOf from 'lodash/_baseIndexOf';
 import isArray from 'lodash/isArray';
 import okeys from 'lodash/keys';
 import noop from 'lodash/noop';
-import slice from 'lodash/slice';
 
+import slice from './internal/slice';
 import once from './internal/once';
 import onlyOnce from './internal/onlyOnce';
 import wrapAsync from './internal/wrapAsync';