--- a/lib/compose.js
+++ b/lib/compose.js
@@ -1,5 +1,5 @@
 import seq from './seq';
-import rest from './internal/rest';
+import slice from './internal/slice';
 
 /**
  * Creates a function which is a composition of the passed asynchronous
@@ -36,6 +36,6 @@ import rest from './internal/rest';
  *     // result now equals 15
  * });
  */
-export default rest(function(args) {
-    return seq.apply(null, args.reverse());
-});
+export default function(/*...args*/) {
+    return seq.apply(null, slice(arguments).reverse());
+};