--- a/lib/constant.js
+++ b/lib/constant.js
@@ -1,5 +1,4 @@
-import rest from './internal/rest';
-import initialParams from './internal/initialParams';
+import slice from './internal/slice';
 
 /**
  * Returns a function that when called, calls-back with the values provided.
@@ -43,9 +42,11 @@ import initialParams from './internal/initialParams';
  *     //...
  * }, callback);
  */
-export default rest(function(values) {
+export default function(/*...values*/) {
+    var values = slice(arguments);
     var args = [null].concat(values);
-    return initialParams(function (ignoredArgs, callback) {
+    return function (/*...ignoredArgs, callback*/) {
+        var callback = arguments[arguments.length - 1];
         return callback.apply(this, args);
-    });
-});
+    };
+};