--- a/lib/doDuring.js
+++ b/lib/doDuring.js
@@ -28,7 +28,7 @@ export default function doDuring(fn, test, callback) {
     var _fn = wrapAsync(fn);
     var _test = wrapAsync(test);
 
-    function next(err/*, args...*/) {
+    function next(err/*, ...args*/) {
         if (err) return callback(err);
         var args = slice(arguments, 1);
         args.push(check);