--- a/lib/doWhilst.js
+++ b/lib/doWhilst.js
@@ -1,5 +1,5 @@
 import noop from 'lodash/noop';
-import rest from './internal/rest';
+import slice from './internal/slice';
 
 import onlyOnce from './internal/onlyOnce';
 import wrapAsync from './internal/wrapAsync';
@@ -29,10 +29,11 @@ import wrapAsync from './internal/wrapAsync';
 export default function doWhilst(iteratee, test, callback) {
     callback = onlyOnce(callback || noop);
     var _iteratee = wrapAsync(iteratee);
-    var next = rest(function(err, args) {
+    var next = function(err/*, args...*/) {
         if (err) return callback(err);
+        var args = slice(arguments, 1);
         if (test.apply(this, args)) return _iteratee(next);
         callback.apply(null, [null].concat(args));
-    });
+    };
     _iteratee(next);
 }