--- a/lib/doWhilst.js
+++ b/lib/doWhilst.js
@@ -29,7 +29,7 @@ import wrapAsync from './internal/wrapAsync';
 export default function doWhilst(iteratee, test, callback) {
     callback = onlyOnce(callback || noop);
     var _iteratee = wrapAsync(iteratee);
-    var next = function(err/*, args...*/) {
+    var next = function(err/*, ...args*/) {
         if (err) return callback(err);
         var args = slice(arguments, 1);
         if (test.apply(this, args)) return _iteratee(next);