--- a/lib/internal/applyEach.js
+++ b/lib/internal/applyEach.js
@@ -1,9 +1,10 @@
-import rest from './rest';
+import slice from './slice';
 import initialParams from './initialParams';
 import wrapAsync from './wrapAsync';
 
 export default function applyEach(eachfn) {
-    return rest(function(fns, args) {
+    return function(fns/*, ...args*/) {
+        var args = slice(arguments, 1);
         var go = initialParams(function(args, callback) {
             var that = this;
             return eachfn(fns, function (fn, cb) {
@@ -16,5 +17,5 @@ export default function applyEach(eachfn) {
         else {
             return go;
         }
-    });
+    };
 }