--- a/lib/internal/consoleFunc.js
+++ b/lib/internal/consoleFunc.js
@@ -1,10 +1,12 @@
 import arrayEach from 'lodash/_arrayEach';
-import rest from './rest';
+import slice from './slice';
 import wrapAsync from './wrapAsync';
 
 export default function consoleFunc(name) {
-    return rest(function (fn, args) {
-        wrapAsync(fn).apply(null, args.concat(rest(function (err, args) {
+    return function (fn/*, ...args*/) {
+        var args = slice(arguments, 1);
+        wrapAsync(fn).apply(null, args.concat(function (err/*, ...args*/) {
+            var args = slice(arguments, 1);
             if (typeof console === 'object') {
                 if (err) {
                     if (console.error) {
@@ -17,6 +19,6 @@ export default function consoleFunc(name) {
                     });
                 }
             }
-        })));
-    });
+        }));
+    };
 }