--- a/lib/internal/consoleFunc.js
+++ b/lib/internal/consoleFunc.js
@@ -5,20 +5,20 @@ import wrapAsync from './wrapAsync';
 export default function consoleFunc(name) {
     return function (fn/*, ...args*/) {
         var args = slice(arguments, 1);
-        wrapAsync(fn).apply(null, args.concat(function (err/*, ...args*/) {
+        args.push(function (err/*, ...args*/) {
             var args = slice(arguments, 1);
             if (typeof console === 'object') {
                 if (err) {
                     if (console.error) {
                         console.error(err);
                     }
-                }
-                else if (console[name]) {
+                } else if (console[name]) {
                     arrayEach(args, function (x) {
                         console[name](x);
                     });
                 }
             }
-        }));
+        })
+        wrapAsync(fn).apply(null, args);
     };
 }