--- a/lib/internal/parallel.js
+++ b/lib/internal/parallel.js
@@ -1,6 +1,6 @@
 import noop from 'lodash/noop';
 import isArrayLike from 'lodash/isArrayLike';
-import rest from './rest';
+import slice from './slice';
 import wrapAsync from './wrapAsync';
 
 export default function _parallel(eachfn, tasks, callback) {
@@ -8,13 +8,13 @@ export default function _parallel(eachfn, tasks, callback) {
     var results = isArrayLike(tasks) ? [] : {};
 
     eachfn(tasks, function (task, key, callback) {
-        wrapAsync(task)(rest(function (err, args) {
-            if (args.length <= 1) {
-                args = args[0];
+        wrapAsync(task)(function (err, result) {
+            if (arguments.length > 2) {
+                result = slice(arguments, 1);
             }
-            results[key] = args;
+            results[key] = result;
             callback(err);
-        }));
+        });
     }, function (err) {
         callback(err, results);
     });