--- a/lib/internal/setImmediate.js
+++ b/lib/internal/setImmediate.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import rest from './rest';
+import slice from './slice';
 
 export var hasSetImmediate = typeof setImmediate === 'function' && setImmediate;
 export var hasNextTick = typeof process === 'object' && typeof process.nextTick === 'function';
@@ -10,11 +10,12 @@ export function fallback(fn) {
 }
 
 export function wrap(defer) {
-    return rest(function (fn, args) {
+    return function (fn/*, ...args*/) {
+        var args = slice(arguments, 1);
         defer(function () {
             fn.apply(null, args);
         });
-    });
+    };
 }
 
 var _defer;