--- a/lib/memoize.js
+++ b/lib/memoize.js
@@ -1,5 +1,5 @@
 import identity from 'lodash/identity';
-import rest from './internal/rest';
+import slice from './internal/slice';
 
 import setImmediate from './internal/setImmediate';
 import initialParams from './internal/initialParams';
@@ -61,14 +61,15 @@ export default function memoize(fn, hasher) {
             queues[key].push(callback);
         } else {
             queues[key] = [callback];
-            _fn.apply(null, args.concat(rest(function(args) {
+            _fn.apply(null, args.concat(function(/*args*/) {
+                var args = slice(arguments);
                 memo[key] = args;
                 var q = queues[key];
                 delete queues[key];
                 for (var i = 0, l = q.length; i < l; i++) {
                     q[i].apply(null, args);
                 }
-            })));
+            }));
         }
     });
     memoized.memo = memo;