--- a/lib/reduceRight.js
+++ b/lib/reduceRight.js
@@ -1,6 +1,5 @@
 import reduce from './reduce';
-
-var slice = Array.prototype.slice;
+import slice from './internal/slice';
 
 /**
  * Same as [`reduce`]{@link module:Collections.reduce}, only operates on `array` in reverse order.
@@ -25,6 +24,6 @@ var slice = Array.prototype.slice;
  * (err, result).
  */
 export default function reduceRight (array, memo, iteratee, callback) {
-    var reversed = slice.call(array).reverse();
+    var reversed = slice(array).reverse();
     reduce(reversed, memo, iteratee, callback);
 }