--- a/lib/reflect.js
+++ b/lib/reflect.js
@@ -1,5 +1,5 @@
 import initialParams from './internal/initialParams';
-import rest from './internal/rest';
+import slice from './internal/slice';
 import wrapAsync from './internal/wrapAsync';
 
 /**
@@ -44,23 +44,19 @@ import wrapAsync from './internal/wrapAsync';
 export default function reflect(fn) {
     var _fn = wrapAsync(fn);
     return initialParams(function reflectOn(args, reflectCallback) {
-        args.push(rest(function callback(err, cbArgs) {
+        args.push(function callback(err, cbArg) {
             if (err) {
                 reflectCallback(null, {
                     error: err
                 });
             } else {
-                var value = null;
-                if (cbArgs.length === 1) {
-                    value = cbArgs[0];
-                } else if (cbArgs.length > 1) {
-                    value = cbArgs;
-                }
+                var value = cbArg;
+                if (arguments.length > 2) value = slice(arguments, 1);
                 reflectCallback(null, {
                     value: value
                 });
             }
-        }));
+        });
 
         return _fn.apply(this, args);
     });