--- a/lib/reflect.js
+++ b/lib/reflect.js
@@ -44,17 +44,17 @@ import wrapAsync from './internal/wrapAsync';
 export default function reflect(fn) {
     var _fn = wrapAsync(fn);
     return initialParams(function reflectOn(args, reflectCallback) {
-        args.push(function callback(err, cbArg) {
-            if (err) {
-                reflectCallback(null, {
-                    error: err
-                });
+        args.push(function callback(error, cbArg) {
+            if (error) {
+                reflectCallback(null, { error: error });
             } else {
-                var value = cbArg;
-                if (arguments.length > 2) value = slice(arguments, 1);
-                reflectCallback(null, {
-                    value: value
-                });
+                var value;
+                if (arguments.length <= 2) {
+                    value = cbArg
+                } else {
+                    value = slice(arguments, 1);
+                }
+                reflectCallback(null, { value: value });
             }
         });