--- a/lib/seq.js
+++ b/lib/seq.js
@@ -1,5 +1,6 @@
 import noop from 'lodash/noop';
 import rest from './internal/rest';
+import slice from 'lodash/slice';
 import reduce from './reduce';
 import wrapAsync from './internal/wrapAsync';
 import arrayMap from 'lodash/_arrayMap';
@@ -55,9 +56,10 @@ export default rest(function seq(functions) {
         }
 
         reduce(_functions, args, function(newargs, fn, cb) {
-            fn.apply(that, newargs.concat(rest(function(err, nextargs) {
+            fn.apply(that, newargs.concat(function(err/*, nextargs...*/) {
+                var nextargs = slice(arguments, 1);
                 cb(err, nextargs);
-            })));
+            }));
         },
         function(err, results) {
             cb.apply(that, [err].concat(results));