--- a/lib/seq.js
+++ b/lib/seq.js
@@ -1,5 +1,4 @@
 import noop from 'lodash/noop';
-import rest from './internal/rest';
 import slice from 'lodash/slice';
 import reduce from './reduce';
 import wrapAsync from './internal/wrapAsync';
@@ -43,9 +42,10 @@ import arrayMap from 'lodash/_arrayMap';
  *     });
  * });
  */
-export default rest(function seq(functions) {
-    var _functions = arrayMap(functions, wrapAsync);
-    return rest(function(args) {
+export default function seq(/*...functions*/) {
+    var _functions = arrayMap(arguments, wrapAsync);
+    return function(/*...args*/) {
+        var args = slice(arguments);
         var that = this;
 
         var cb = args[args.length - 1];
@@ -56,7 +56,7 @@ export default rest(function seq(functions) {
         }
 
         reduce(_functions, args, function(newargs, fn, cb) {
-            fn.apply(that, newargs.concat(function(err/*, nextargs...*/) {
+            fn.apply(that, newargs.concat(function(err/*, ...nextargs*/) {
                 var nextargs = slice(arguments, 1);
                 cb(err, nextargs);
             }));
@@ -64,5 +64,5 @@ export default rest(function seq(functions) {
         function(err, results) {
             cb.apply(that, [err].concat(results));
         });
-    });
-})
+    };
+}