--- a/lib/seq.js
+++ b/lib/seq.js
@@ -1,5 +1,5 @@
 import noop from 'lodash/noop';
-import slice from 'lodash/slice';
+import slice from './internal/slice';
 import reduce from './reduce';
 import wrapAsync from './internal/wrapAsync';
 import arrayMap from 'lodash/_arrayMap';