--- a/lib/waterfall.js
+++ b/lib/waterfall.js
@@ -2,6 +2,7 @@ import isArray from 'lodash/isArray';
 import noop from 'lodash/noop';
 import once from './internal/once';
 import rest from './internal/rest';
+import slice from './internal/slice';
 
 import onlyOnce from './internal/onlyOnce';
 import wrapAsync from './internal/wrapAsync';
@@ -74,12 +75,13 @@ export default  function(tasks, callback) {
             return callback.apply(null, [null].concat(args));
         }
 
-        var taskCallback = onlyOnce(rest(function(err, args) {
+        var taskCallback = onlyOnce(function(err/*, cbArgs...*/) {
+            var cbArgs = slice(arguments, 1);
             if (err) {
-                return callback.apply(null, [err].concat(args));
+                return callback.apply(null, [err].concat(cbArgs));
             }
-            nextTask(args);
-        }));
+            nextTask(cbArgs);
+        });
 
         args.push(taskCallback);