--- a/lib/waterfall.js
+++ b/lib/waterfall.js
@@ -1,7 +1,6 @@
 import isArray from 'lodash/isArray';
 import noop from 'lodash/noop';
 import once from './internal/once';
-import rest from './internal/rest';
 import slice from './internal/slice';
 
 import onlyOnce from './internal/onlyOnce';