--- a/lib/waterfall.js
+++ b/lib/waterfall.js
@@ -68,25 +68,23 @@ export default  function(tasks, callback) {
     if (!isArray(tasks)) return callback(new Error('First argument to waterfall must be an array of functions'));
     if (!tasks.length) return callback();
     var taskIndex = 0;
+    var args = [];
 
-    function nextTask(args) {
-        if (taskIndex === tasks.length) {
-            return callback.apply(null, [null].concat(args));
-        }
-
-        var taskCallback = onlyOnce(function(err/*, cbArgs...*/) {
-            var cbArgs = slice(arguments, 1);
-            if (err) {
-                return callback.apply(null, [err].concat(cbArgs));
-            }
-            nextTask(cbArgs);
-        });
-
+    function nextTask() {
+        var task = wrapAsync(tasks[taskIndex++]);
+        var taskCallback = onlyOnce(next);
         args.push(taskCallback);
 
-        var task = wrapAsync(tasks[taskIndex++]);
         task.apply(null, args);
     }
 
-    nextTask([]);
+    function next(err/*, ...args*/) {
+        if (err || taskIndex === tasks.length) {
+            return callback.apply(null, arguments);
+        }
+        args = slice(arguments, 1);
+        nextTask();
+    }
+
+    nextTask();
 }