--- a/lib/waterfall.js
+++ b/lib/waterfall.js
@@ -68,13 +68,10 @@ export default  function(tasks, callback) {
     if (!isArray(tasks)) return callback(new Error('First argument to waterfall must be an array of functions'));
     if (!tasks.length) return callback();
     var taskIndex = 0;
-    var args = [];
 
-    function nextTask() {
+    function nextTask(args) {
         var task = wrapAsync(tasks[taskIndex++]);
-        var taskCallback = onlyOnce(next);
-        args.push(taskCallback);
-
+        args.push(onlyOnce(next));
         task.apply(null, args);
     }
 
@@ -82,9 +79,8 @@ export default  function(tasks, callback) {
         if (err || taskIndex === tasks.length) {
             return callback.apply(null, arguments);
         }
-        args = slice(arguments, 1);
-        nextTask();
+        nextTask(slice(arguments, 1));
     }
 
-    nextTask();
+    nextTask([]);
 }