--- a/lib/whilst.js
+++ b/lib/whilst.js
@@ -1,5 +1,5 @@
 import noop from 'lodash/noop';
-import rest from './internal/rest';
+import slice from './internal/slice';
 
 import onlyOnce from './internal/onlyOnce';
 import wrapAsync from './internal/wrapAsync';
@@ -42,10 +42,10 @@ export default function whilst(test, iteratee, callback) {
     callback = onlyOnce(callback || noop);
     var _iteratee = wrapAsync(iteratee);
     if (!test()) return callback(null);
-    var next = rest(function(err, args) {
+    var next = function(err/*, args...*/) {
         if (err) return callback(err);
         if (test()) return _iteratee(next);
-        callback.apply(null, [null].concat(args));
-    });
+        callback.apply(null, [null].concat(slice(arguments, 1)));
+    };
     _iteratee(next);
 }