--- a/lib/whilst.js
+++ b/lib/whilst.js
@@ -42,7 +42,7 @@ export default function whilst(test, iteratee, callback) {
     callback = onlyOnce(callback || noop);
     var _iteratee = wrapAsync(iteratee);
     if (!test()) return callback(null);
-    var next = function(err/*, args...*/) {
+    var next = function(err/*, ...args*/) {
         if (err) return callback(err);
         if (test()) return _iteratee(next);
         callback.apply(null, [null].concat(slice(arguments, 1)));