--- a/lib/whilst.js
+++ b/lib/whilst.js
@@ -45,7 +45,8 @@ export default function whilst(test, iteratee, callback) {
     var next = function(err/*, ...args*/) {
         if (err) return callback(err);
         if (test()) return _iteratee(next);
-        callback.apply(null, [null].concat(slice(arguments, 1)));
+        var args = slice(arguments, 1);
+        callback.apply(null, [null].concat(args));
     };
     _iteratee(next);
 }