I ran some profiling, and notices that a lot of time is being spent in `internal/rest`.  It turns out `rest()` is much slower than a straightforward arguments `slice()`.  (I also tried the "inline slice" strategy with a loop that is sometimes recommended, it made absolutely no difference).  The differences are significant, especially for `parallel`:

<details>
   <summary>Benchmark Results (click to expand)</summary>
<pre>
$ node perf/benchmark.js -g 'series|waterfall|auto|parallel'
Latest tag is  v2.3.0
Comparing v2.3.0 with current on Node v7.7.4
--------------------------------------
(node:29289) DeprecationWarning: Calling an asynchronous function without callback is deprecated.
eachSeries(10) v2.3.0 x 51,744 ops/sec ±1.47% (27 runs sampled), 0.0193ms per run
eachSeries(10) current x 54,845 ops/sec ±1.50% (29 runs sampled), 0.0182ms per run
current is faster
--------------------------------------
eachSeries(300) v2.3.0 x 2,718 ops/sec ±2.76% (29 runs sampled), 0.368ms per run
eachSeries(300) current x 3,108 ops/sec ±1.69% (25 runs sampled), 0.322ms per run
current is faster
--------------------------------------
eachSeries(10000) v2.3.0 x 87.15 ops/sec ±2.30% (29 runs sampled), 11.5ms per run
eachSeries(10000) current x 106 ops/sec ±0.96% (30 runs sampled), 9.42ms per run
current is faster
--------------------------------------
mapSeries(10) v2.3.0 x 51,520 ops/sec ±1.06% (31 runs sampled), 0.0194ms per run
mapSeries(10) current x 55,203 ops/sec ±1.58% (30 runs sampled), 0.0181ms per run
current is faster
--------------------------------------
mapSeries(300) v2.3.0 x 2,832 ops/sec ±0.96% (30 runs sampled), 0.353ms per run
mapSeries(300) current x 3,108 ops/sec ±1.62% (32 runs sampled), 0.322ms per run
current is faster
--------------------------------------
mapSeries(10000) v2.3.0 x 85.19 ops/sec ±2.45% (29 runs sampled), 11.7ms per run
mapSeries(10000) current x 95.28 ops/sec ±2.42% (27 runs sampled), 10.5ms per run
current is faster
--------------------------------------
eachOfSeries(10) v2.3.0 x 55,509 ops/sec ±1.51% (30 runs sampled), 0.0180ms per run
eachOfSeries(10) current x 58,832 ops/sec ±2.64% (29 runs sampled), 0.0170ms per run
current is faster
--------------------------------------
eachOfSeries(300) v2.3.0 x 3,000 ops/sec ±1.96% (28 runs sampled), 0.333ms per run
eachOfSeries(300) current x 3,369 ops/sec ±1.01% (31 runs sampled), 0.297ms per run
current is faster
--------------------------------------
eachOfSeries(10000) v2.3.0 x 93.05 ops/sec ±1.88% (32 runs sampled), 10.7ms per run
eachOfSeries(10000) current x 103 ops/sec ±2.38% (30 runs sampled), 9.67ms per run
current is faster
--------------------------------------
parallel(10) v2.3.0 x 79,183 ops/sec ±1.84% (29 runs sampled), 0.0126ms per run
parallel(10) current x 93,866 ops/sec ±2.27% (30 runs sampled), 0.0107ms per run
current is faster
--------------------------------------
parallel(100) v2.3.0 x 17,376 ops/sec ±1.95% (30 runs sampled), 0.0576ms per run
parallel(100) current x 25,385 ops/sec ±2.02% (29 runs sampled), 0.0394ms per run
current is faster
--------------------------------------
parallel(1000) v2.3.0 x 1,952 ops/sec ±1.56% (29 runs sampled), 0.512ms per run
parallel(1000) current x 3,011 ops/sec ±2.09% (29 runs sampled), 0.332ms per run
current is faster
--------------------------------------
series(10) v2.3.0 x 48,445 ops/sec ±2.92% (24 runs sampled), 0.0206ms per run
series(10) current x 57,840 ops/sec ±1.63% (30 runs sampled), 0.0173ms per run
current is faster
--------------------------------------
series(100) v2.3.0 x 7,491 ops/sec ±1.65% (30 runs sampled), 0.133ms per run
series(100) current x 8,525 ops/sec ±2.66% (28 runs sampled), 0.117ms per run
current is faster
--------------------------------------
series(1000) v2.3.0 x 814 ops/sec ±2.15% (31 runs sampled), 1.23ms per run
series(1000) current x 957 ops/sec ±2.25% (30 runs sampled), 1.05ms per run
current is faster
--------------------------------------
waterfall(10) v2.3.0 x 42,529 ops/sec ±1.92% (30 runs sampled), 0.0235ms per run
waterfall(10) current x 47,484 ops/sec ±2.51% (30 runs sampled), 0.0211ms per run
current is faster
--------------------------------------
waterfall(100) v2.3.0 x 6,180 ops/sec ±2.59% (31 runs sampled), 0.162ms per run
waterfall(100) current x 7,322 ops/sec ±1.33% (27 runs sampled), 0.137ms per run
current is faster
--------------------------------------
waterfall(1000) v2.3.0 x 656 ops/sec ±1.26% (31 runs sampled), 1.52ms per run
waterfall(1000) current x 771 ops/sec ±1.50% (30 runs sampled), 1.30ms per run
current is faster
--------------------------------------
auto(5) v2.3.0 x 36,391 ops/sec ±2.34% (30 runs sampled), 0.0275ms per run
auto(5) current x 39,932 ops/sec ±2.13% (29 runs sampled), 0.0250ms per run
current is faster
--------------------------------------
auto(10) v2.3.0 x 20,148 ops/sec ±2.08% (29 runs sampled), 0.0496ms per run
auto(10) current x 21,765 ops/sec ±1.94% (29 runs sampled), 0.0459ms per run
current is faster
--------------------------------------
auto(100) v2.3.0 x 866 ops/sec ±1.97% (30 runs sampled), 1.16ms per run
auto(100) current x 903 ops/sec ±1.99% (29 runs sampled), 1.11ms per run
current is faster
--------------------------------------
current faster overall (34.8ms total vs. 40ms total)
current won more benchmarks (21 vs. 0)

</pre>
</details>

<br>

We're also competitive with Bluebird again:

<details>
<summary>Bluebird benchmarks (click to expand)</summary>
<pre>
results for 10000 parallel executions, 1 ms per I/O op

file                                      time(ms)  memory(MB)
callbacks-baseline.js                          247       71.74
callbacks-suguru03-neo-async-parallel.js       369       89.66
callbacks-caolan-async-parallel.js             500      122.13
promises-bluebird-generator.js                 528      105.73
promises-lvivski-davy.js                       554      163.70
promises-bluebird.js                           620      103.58
promises-cujojs-when.js                        736      166.95
//...

Platform info:
Linux 4.8.0-41-generic x64
Node.JS 7.7.4
V8 5.5.372.42
Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz × 8

--------

results for 10000 parallel executions, 1 ms per I/O op

file                                       time(ms)  memory(MB)
callbacks-baseline.js                           113       26.87
callbacks-suguru03-neo-async-waterfall.js       147       36.47
promises-bluebird-generator.js                  211       37.81
callbacks-caolan-async-waterfall.js             222       46.56
promises-bluebird.js                            247       47.08
promises-then-promise.js                        387       86.44
promises-lvivski-davy.js                        399       99.09
promises-cujojs-when.js                         403       72.13
promises-tildeio-rsvp.js                        415       78.18
promises-dfilatov-vow.js                        614      134.69
promises-calvinmetcalf-lie.js                   677      162.16
generators-tj-co.js                             713      112.79
//...

Platform info:
Linux 4.8.0-41-generic x64
Node.JS 7.7.4
V8 5.5.372.42
Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz × 8

</pre>
</details>

<br>

It might be worth getting rid of `rest()` completely, but the gains will be less.

Nothing beats native `...rest` params in ES6, though...

```
$ node perf/benchmark.js -g 'waterfall'
Latest tag is  v2.3.0
Comparing v2.3.0 with current on Node v7.7.4
--------------------------------------
waterfall(10) v2.3.0 x 41,937 ops/sec ±1.54% (23 runs sampled), 0.0238ms per run
waterfall(10) current x 56,884 ops/sec ±2.12% (29 runs sampled), 0.0176ms per run
current is faster
--------------------------------------
waterfall(100) v2.3.0 x 6,061 ops/sec ±2.00% (30 runs sampled), 0.165ms per run
waterfall(100) current x 8,876 ops/sec ±2.65% (28 runs sampled), 0.113ms per run
current is faster
--------------------------------------
waterfall(1000) v2.3.0 x 638 ops/sec ±3.13% (29 runs sampled), 1.57ms per run
waterfall(1000) current x 1,024 ops/sec ±2.06% (30 runs sampled), 0.977ms per run
current is faster
--------------------------------------
current faster overall (1.11ms total vs. 1.76ms total)
current won more benchmarks (3 vs. 0)

-----------

results for 10000 parallel executions, 1 ms per I/O op

file                                       time(ms)  memory(MB)
callbacks-baseline.js                           118       24.98
callbacks-caolan-async-waterfall.js             202       45.82
callbacks-suguru03-neo-async-waterfall.js       203       36.24
promises-bluebird-generator.js                  218       37.35
promises-bluebird.js                            264       53.38

```