@megawac 2017-04-03
Unless JavaScript engines have changed how they handle arguments I question these results and suggest you test in a couple different environments (different versions of node, different versions of chrome, firefox, safari, etc). 

@jdalton, @jridgewell and I investigated this pretty closely about 2 years ago. Reference ticket https://github.com/jashkenas/underscore/pull/2140.