@megawac 2017-04-03
Also we changed from `slice` to `rest` about a year ago after observing similar results when profiling