@hargasinski 2017-04-04
@megawac v8, at least with `TurboFan`, appears to be optimizing the `arguments` object. There's [this design doc](https://docs.google.com/document/d/1DvDx3Xursn1ViV5k4rT4KB8HBfBb2GdUy3wzNfJWcKM/edit#heading=h.xtyvxc2lhjoi) from [their wiki](https://github.com/v8/v8/wiki/TurboFan) that goes a little into the detail of `TurboFan`'s vs `CrankShaft`'s handling of `arguments` and what they are working on.

From what I understand from [petkaantonov's Optimization killers](https://github.com/petkaantonov/bluebird/wiki/Optimization-killers#32-leaking-arguments) and [what the design doc says about `CrankShaft`](https://docs.google.com/document/d/1DvDx3Xursn1ViV5k4rT4KB8HBfBb2GdUy3wzNfJWcKM/edit#heading=h.q5pqrapefnqa), this used to be deoptimized, due to passing `arguments` to another function while [`overRest`'s implementation](https://github.com/lodash/lodash/blob/4.17.4/lodash.js#L6583) creates an Array inline. If v8 now

> always materializes the objects right after the prologue if they’re being used in the function.

then using `slice` with `arguments` is a quick substitute for a rest parameter, and we don't need the other functionality that `overRest` provides.

But yes, this will (most likely) be slower in other browsers/older versions of Node.