@aearly 2017-04-04
I'm seeing a similar performance gain with node 0.12:

```
Latest tag is  v2.3.0
Comparing v2.3.0 with current on Node v0.12.18
--------------------------------------
parallel(10) v2.3.0 x 44,422 ops/sec ±2.65% (28 runs sampled), 0.0225ms per run
parallel(10) current x 53,069 ops/sec ±3.36% (29 runs sampled), 0.0188ms per run
current is faster
--------------------------------------
parallel(100) v2.3.0 x 11,965 ops/sec ±0.61% (32 runs sampled), 0.0836ms per run
parallel(100) current x 18,318 ops/sec ±3.63% (30 runs sampled), 0.0546ms per run
current is faster
--------------------------------------
parallel(1000) v2.3.0 x 1,375 ops/sec ±0.74% (31 runs sampled), 0.727ms per run
parallel(1000) current x 2,430 ops/sec ±3.29% (29 runs sampled), 0.412ms per run
current is faster
--------------------------------------
waterfall(10) v2.3.0 x 28,799 ops/sec ±1.66% (30 runs sampled), 0.0347ms per run
waterfall(10) current x 31,131 ops/sec ±4.33% (29 runs sampled), 0.0321ms per run
current is faster
--------------------------------------
waterfall(100) v2.3.0 x 4,649 ops/sec ±1.48% (31 runs sampled), 0.215ms per run
waterfall(100) current x 5,377 ops/sec ±1.88% (29 runs sampled), 0.186ms per run
current is faster
--------------------------------------
waterfall(1000) v2.3.0 x 492 ops/sec ±1.51% (32 runs sampled), 2.03ms per run
waterfall(1000) current x 580 ops/sec ±1.41% (31 runs sampled), 1.73ms per run
current is faster
--------------------------------------
current faster overall (2.43ms total vs. 3.12ms total)
current won more benchmarks (6 vs. 0)
```
(Benchmarks don't work in node 0.10).  Currently working on a browser benchmark.

I looked at the benchmarks associated with that underscore ticket.  The `slice` implementation uses `arrayProto.slice.call(arguments)` which will always be slower due to `.call()`.  I also notice there is a bug in jridgewell's `for`-loop based implementation -- in `.trigger()` he is essentially doing `arguments.slice(0)` when it should be `arguments.slice(1)`, which could be making slower.