@megawac 2017-04-04
In that case, there were also benchmarks created for lodash which was previously using `_.slice` before `restParam` was added; in that case IIRC the perf gain was slight but the biggest gain was readability.