@megawac 2017-04-04
> then using slice with arguments is a quick substitute for a rest parameter, and we don't need the other functionality that overRest provides.
> But yes, this will (most likely) be slower in other browsers/older versions of Node.

I haven't looked into this in a while but it used to be debilitating in Firefox, okay in Chrome and IE, and pretty bad in Opera. Please make some browser tests before merging this