@aearly 2017-04-04
We started using `rest` between v1.3.0 and v1.4.0.

<details><summary> Node 7 Benchmarks </summary><p>
<pre>
$ node perf/benchmark.js -g 'waterfall|parallel' v1.3.0 v1.4.0
Comparing v1.3.0 with v1.4.0 on Node v7.7.4
--------------------------------------
parallel(10) v1.3.0 x 95,854 ops/sec ±0.91% (27 runs sampled), 0.0104ms per run
parallel(10) v1.4.0 x 84,751 ops/sec ±2.31% (27 runs sampled), 0.0118ms per run
v1.3.0 is faster
--------------------------------------
parallel(100) v1.3.0 x 24,968 ops/sec ±1.95% (27 runs sampled), 0.0401ms per run
parallel(100) v1.4.0 x 20,407 ops/sec ±1.51% (28 runs sampled), 0.0490ms per run
v1.3.0 is faster
--------------------------------------
parallel(1000) v1.3.0 x 3,037 ops/sec ±1.52% (28 runs sampled), 0.329ms per run
parallel(1000) v1.4.0 x 2,303 ops/sec ±1.63% (29 runs sampled), 0.434ms per run
v1.3.0 is faster
--------------------------------------
waterfall(10) v1.3.0 x 19,206 ops/sec ±3.37% (24 runs sampled), 0.0521ms per run
waterfall(10) v1.4.0 x 18,883 ops/sec ±2.88% (27 runs sampled), 0.0530ms per run
Tie
--------------------------------------
waterfall(100) v1.3.0 x 2,589 ops/sec ±2.26% (29 runs sampled), 0.386ms per run
waterfall(100) v1.4.0 x 2,427 ops/sec ±3.31% (28 runs sampled), 0.412ms per run
v1.3.0 is faster
--------------------------------------
waterfall(1000) v1.3.0 x 271 ops/sec ±3.85% (28 runs sampled), 3.69ms per run
waterfall(1000) v1.4.0 x 243 ops/sec ±2.90% (27 runs sampled), 4.11ms per run
v1.3.0 is faster
--------------------------------------
v1.3.0 faster overall (4.51ms total vs. 5.07ms total)
v1.3.0 won more benchmarks (5 vs. 0)
</pre>
</p></details><br>

<details><summary> Node 0.10 Benchmarks </summary><p>
<pre>
$ node perf/benchmark.js -g 'waterfall|parallel' v1.3.0 v1.4.0
Comparing v1.3.0 with v1.4.0 on Node v0.10.48
--------------------------------------
parallel(10) v1.3.0 x 36,982 ops/sec ±3.28% (26 runs sampled), 0.0270ms per run
parallel(10) v1.4.0 x 34,188 ops/sec ±2.63% (26 runs sampled), 0.0293ms per run
v1.3.0 is faster
--------------------------------------
parallel(100) v1.3.0 x 9,111 ops/sec ±2.07% (30 runs sampled), 0.110ms per run
parallel(100) v1.4.0 x 7,572 ops/sec ±2.34% (27 runs sampled), 0.132ms per run
v1.3.0 is faster
--------------------------------------
parallel(1000) v1.3.0 x 1,051 ops/sec ±3.07% (27 runs sampled), 0.952ms per run
parallel(1000) v1.4.0 x 782 ops/sec ±3.82% (28 runs sampled), 1.28ms per run
v1.3.0 is faster
--------------------------------------
waterfall(10) v1.3.0 x 20,187 ops/sec ±2.54% (27 runs sampled), 0.0495ms per run
waterfall(10) v1.4.0 x 17,901 ops/sec ±2.79% (26 runs sampled), 0.0559ms per run
v1.3.0 is faster
--------------------------------------
waterfall(100) v1.3.0 x 2,892 ops/sec ±2.27% (25 runs sampled), 0.346ms per run
waterfall(100) v1.4.0 x 2,662 ops/sec ±1.64% (27 runs sampled), 0.376ms per run
v1.3.0 is faster
--------------------------------------
waterfall(1000) v1.3.0 x 315 ops/sec ±2.22% (31 runs sampled), 3.17ms per run
waterfall(1000) v1.4.0 x 285 ops/sec ±3.53% (25 runs sampled), 3.51ms per run
v1.3.0 is faster
--------------------------------------
v1.3.0 faster overall (4.66ms total vs. 5.38ms total)
v1.3.0 won more benchmarks (6 vs. 0)
</pre>

</p></details><br>

It was a 10-12% performance regression.  This PR uses a more streamlined version of slice, so the gains are a bit more.

Still working on the browser benchmark, jsperf.com is being a PITA right now....